+++ tests/test_raster.py
@@ -9,7 +9,7 @@ from shapely.geometry import box, Point, LineString
 import dask
 from osgeo import gdal
 from affine import Affine
-from rasterio.transform import array_bounds
+import os
 
 from hydromt import open_raster, gis_utils
 from hydromt import raster
@@ -429,3 +429,43 @@ def test_rotated(transform, shape, tmpdir):
     dst_crs = gis_utils.parse_crs("utm", da.raster.bounds)
     da2_reproj = da2.raster.reproject(dst_crs=dst_crs)
     assert np.all(da2.raster.box.intersects(da2_reproj.raster.box.to_crs(4326)))
+
+
+def test_to_xyz_tiles(tmpdir, rioda_large):
+    # NOTE: this method does not work in debug mode because of os.subprocess
+    path = str(tmpdir)
+    rioda_large.raster.to_xyz_tiles(
+        os.path.join(path, "dummy_xyz"),
+        tile_size=256,
+        zoom_levels=[0, 2],
+    )
+    with open(os.path.join(path, "dummy_xyz", "0", "filelist.txt"), "r") as f:
+        assert len(f.readlines()) == 16
+    with open(os.path.join(path, "dummy_xyz", "2", "filelist.txt"), "r") as f:
+        assert len(f.readlines()) == 1
+
+    # test create_vrt
+    vrt_fn = os.path.join(path, "dummy_xyz", "vrt", "zl0.vrt")
+    files_path = os.path.join(path, "dummy_xyz", "*", "*", "*.tif")
+    gis_utils.create_vrt(vrt_fn, files_path=files_path)
+    assert os.path.isfile(vrt_fn)
+    assert isinstance(open_raster(vrt_fn).load(), xr.DataArray)  # try reading
+    with pytest.raises(
+        ValueError, match="Either 'file_list_path' or 'files_path' is required"
+    ):
+        gis_utils.create_vrt(vrt_fn)
+    with pytest.raises(IOError, match="No files found at "):
+        gis_utils.create_vrt(
+            vrt_fn, files_path=os.path.join(path, "dummy_xyz", "*.abc")
+        )
+
+
+# def test_to_osm(tmpdir, dummy):
+#     path = str(tmpdir)
+#     dummy.raster.to_osm(
+#         f"{path}\\dummy_osm",
+#         zl=4,
+#         bbox=(0, -45, 45, 0),
+#     )
+#     f = open(f"{path}\\dummy_osm\\3\\filelist.txt", "r")
+#     assert len(f.readlines()) == 4